mainrepo/lib/dl_configs/pyproject.toml

[tool.poetry]
name = "datalens-configs"
version = "0.0.1"
description = ""
authors = ["Konstantin Chupin <kchupin@yandex-team.ru>"]
packages = [{include = "dl_configs"}]
license = "MIT"


[tool.poetry.dependencies]
attrs = ">=22.2.0"
cryptography = ">=3.3.2"
dynamic-enum = {path = "../dynamic_enum"}
python = ">=3.10, <3.12"
pyyaml = ">=5.3.1"
typeguard = ">=2.13.3"
datalens-utils = {path = "../dl_utils"}
datalens-constants = {path = "../dl_constants"}

[tool.poetry.group]
[tool.poetry.group.tests.dependencies]
shortuuid = ">=1.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra"
testpaths = ["dl_configs_tests/unit"]



[tool.mypy]
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_optional = true
